+++ packages/dian/src/Soap/Responses/UploadDocumentMessage.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tribux\Dian\Soap\Responses;
+
+final readonly class UploadDocumentMessage
+{
+    public function __construct(
+        public ?string $documentKey,
+        public ?string $processedMessage,
+        public ?string $senderCode,
+        public ?bool $success,
+        public ?string $xmlFileName,
+    ) {
+    }
+}